+++ CHANGELOG.md
@@ -74,6 +74,7 @@ tech changes will usually be stripped from release notes for the public
 -   Moving polygon point when polyon is rotated
 -   Normal shapes sometimes wrongly being identified as spawn locations
 -   Dice sometimes going offscreen
+-   Keyboard movement not moving select-tool's rotation helper
 -   [DM] Assets not being able to moved up to parent folder
 -   [DM] Assets not being removable if a shape with a link to the asset exists
 -   [DM] Annotations still being visible until refresh after removing player access